fix loop by separating controller and allow asserting without update
