+++ src/Service/TransactionFlow/SaferPayTransactionAssertion.php
@@ -66,9 +66,10 @@ class SaferPayTransactionAssertion
      * @return AssertBody
      * @throws \Exception
      */
-    public function assert($cartId)
+    public function assert($cartId, $update = true)
     {
         $saferPayOrder = new SaferPayOrder($this->orderRepository->getIdByCartId($cartId));
+        \PrestaShopLogger::addLog('saferpayOrderId:' . $saferPayOrder->id);
 
         $assertRequest = $this->assertRequestCreator->create($saferPayOrder->token);
         $assertResponse = $this->assertionService->assert($assertRequest, $saferPayOrder->id);
@@ -82,9 +83,12 @@ class SaferPayTransactionAssertion
             $saferPayOrder->id
         );
 
-        $saferPayOrder->transaction_id = $assertBody->getTransaction()->getId();
-        $saferPayOrder->id_cart = $cartId;
-        $saferPayOrder->update();
+        // assertion shouldn't update, this is quickfix for what seems to be a general flaw in structure
+        if ($update) {
+            $saferPayOrder->transaction_id = $assertBody->getTransaction()->getId();
+            $saferPayOrder->id_cart = $cartId;
+            $saferPayOrder->update();
+        }
 
         return $assertBody;
     }
